Initial work on network types
