+++ stdlib/crochet.network.types/source/internal.crochet
@@ -1,0 +1,7 @@
+% crochet
+
+singleton internal;
+capability internal;
+
+protect type internal with internal;
+protect global internal with internal;
